Make Coin its own class
Instead of just a u64, which is converted to BigNum in wasm, we have our
own wrapper type similar to js-chain-libs.

The reason for this is that BigNUm might not exist in all frameworks
(e.g. ReactNative) so we need to have the ability to create/get them as
strings.