Clarify u32 vs  u64 usage for CDDL 'uints'
Metadata needys to be U64, the encoding of metadata contents needed u64
as well.

Epoch/slot/tx index/cert index are fine as u32 as it will be many
hundreds of thousands or more of years until that is an issue.

prelude.rs was also split up into error.rs and the rest into utils.rs
because half of it was either hand-written or hand-ededit from
cddl-codgen and the file itself is a static file within cddl-codgen so
there shouldn't be any changes as a result of re-generation of the cddl.