make txbuilder compile
add minimum_utxo_val method and rename the input argument

for now it only uses the argument given just as before, as I don't see
any point in hardcoding it when it wasn't before. Probably the `new`
constructor should take the three constants needed to do the proper
computation.

returns an error when outputs include native assets, as the computation
is different and is not really implemented