Change existing structs for Alonzo
Only the core data model was updated.

Auxiliary functionality like tx_builder was only changed to compile, so
it might need changing to take advantage of new functionality.

Test vectors for fees and such were not changed and are possibly wrong
if the underlying fee model has changed from Shelley-Mary to Alonzo.